model/requirements.txt
pandas
scikit-learn
joblib
google-cloud-storage
google-cloud-bigquery